fixed crash in fft

+++ semestr-1/podstawy-informatyki/lab-a4/main.py
@@ -26,7 +26,7 @@ while True:
 
     while True: # Timespan
         try:
-            timespan = float(input("Podaj czas trwania sygnału [s]:"))
+            timespan = int(input("Podaj czas trwania sygnału [s]:"))
 
             if timespan <= 0:
                 raise ValueError
